Minor changes to ImagesListPresenter and ImagesHelper
